pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "riplex"
dynamic = ["version"]
description = "Automates the tedious manual work around MakeMKV: figuring out what to rip, which MKV files are actually what, and organizing everything into Plex-compatible folder structures."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27",
    "dvdcompare-scraper>=0.1.6",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]
gui = [
    "flet>=0.84",
]

[project.scripts]
riplex = "riplex_cli.main:main"

[project.gui-scripts]
riplex-ui = "riplex_app.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[tool.pytest.ini_options]
testpaths = ["tests"]
